refactor: Add support for removing files from LSH index
The code changes introduce a new method in the LSH class to remove a file's hash from the LSH index. This is useful for updating the index when a similar file is replaced during media processing. The commit message follows the established convention of using a prefix to indicate the type of change (refactor for code refactoring).